@@ -1,8 +1,6 @@
 package sample.results.test.multiple.bitrise.com.multipletestresultssample;
 
-import android.content.Context;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.runner.AndroidJUnit4;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -55,45 +53,5 @@ public class InstrumentedTest0 {
     @Test
     public void correctUITestCase9() throws Exception {
         assertEquals(4, 2+2);
-    }
-    @Test
-    public void failedUITestCase0() throws Exception {
-        assertEquals(4, 2+1);
-    }
-    @Test
-    public void failedUITestCase1() throws Exception {
-        assertEquals(4, 2+1);
-    }
-    @Test
-    public void failedUITestCase2() throws Exception {
-        assertEquals(4, 2+1);
-    }
-    @Test
-    public void failedUITestCase3() throws Exception {
-        assertEquals(4, 2+1);
-    }
-    @Test
-    public void failedUITestCase4() throws Exception {
-        assertEquals(4, 2+1);
-    }
-    @Test
-    public void failedUITestCase5() throws Exception {
-        assertEquals(4, 2+1);
-    }
-    @Test
-    public void failedUITestCase6() throws Exception {
-        assertEquals(4, 2+1);
-    }
-    @Test
-    public void failedUITestCase7() throws Exception {
-        assertEquals(4, 2+1);
-    }
-    @Test
-    public void failedUITestCase8() throws Exception {
-        assertEquals(4, 2+1);
-    }
-    @Test
-    public void failedUITestCase9() throws Exception {
-        assertEquals(4, 2+1);
     }
 }
